make AddressView a separated file. add AddressView to transactionHeaderView to address sec-review. add zero-width space to make sure account full address will not truncate in context menu.
